CHANGE: Moves ImmutableCommand to a more logical location.
